bracket components now displaying usernames by round correctly
